--- src/components/pages/About.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function About() {
-    return (
-        <div>
-            <h1>This is the About page!</h1>
-        </div>
-    )
-}
