--- src/components/pages/CharacterPage.js
@@ -1,10 +1,0 @@
-import React from 'react'
-
-export default function CharacterPage(props) {
-    console.log(props)
-    return (
-        <div>
-           Hello
-        </div>
-    )
-}
