--- src/components/pages/Combo.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function Combos() {
-    return (
-        <div>
-            <h1>This is the Combo page!</h1>
-        </div>
-    )
-}
