--- src/components/pages/User.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function User() {
-    return (
-        <div>
-            <h1>This is the User page!</h1>
-        </div>
-    )
-}
