+++ src/index.js
@@ -4,9 +4,4 @@ import App from './App';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+ReactDOM.render(<App />, document.getElementById('root'));
