+++ src/pages/CharacterPage.js
@@ -1,0 +1,10 @@
+import React from 'react'
+
+export default function CharacterPage(props) {
+  console.log(props);
+  return (
+    <div>
+      Hello
+    </div>
+  );
+}
