+++ src/pages/Profile.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Profile() {
+  return (
+    <div>
+      <h1>This is the user profile page!</h1>
+    </div>
+  );
+}
