+++ src/pages/Resources.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default function Resources () {
+  return (
+    <div>
+      <h1>This is the About page!</h1>
+    </div>
+  );
+}
